Beginnings of an AST
Basic `Node` definition, as well as the idea of an environment. It's
possible that the environment should just be another node, but that is
something to decide at some other time, I think.

I also think we will need to keep track of all the lexed tokens so we
can look backwards, a simple linked list should suffice. We can always
upgrade to a hashmap with relative ease, really.