Progress made on-stream on the 10th of August, 2022
We have removed type-checking from the parsing stage, and simply pass
through parsed expressions during initialization/assignment directly,
unchanged and unchecked. This means we will have to create a
typechecking stage in between parsing and codegen which ensures that
the returned value of expressions matches the use case.

Also in this commit: very (very) basic codegen!
We now have the mocked out files and functions to begin work on actual
code generation. The parsing context/environment is now used
differently, more in accordance with what is needed for typechecking
and codegen.

All in all, we didn't make a ton of progress per-se, but we did have a
fun, positive time and were able to make some progress, so I'd say it
was worth it.