Parser: Function Calls (and more)
Function calls are now parsed properly! Hurray!

`main.c` was cleaned up greatly, as the parsing of expressions into a
singular program has moved to a helper within the parser,
`parse_program()`. With this done, we are actually beginning to get
somewhere, and if we were to add arithmetic and logical nodes that
operated on integers, everything would start to fall into place.