!Overhaul CodeGen!
Codegen was unnecessarily messy and a terrible API, really, that would
not have even worked for basic math expressions. This new system
allows for all that and more :^).

Shout-out to Douglas Thain for making the following video; it helped
me understand code generation and what problems need solved.
https://www.youtube.com/watch?v=-ti07Z0xKKg

Shout-out to Anita R for making the following video: it details the
overall structure of a compiler and specifically codegen issues.
https://www.youtube.com/watch?v=0SagYu6C0d4