Progress made on-stream on the 19th of August, 2022
We were able to finally fix that bug with not parsing a variable
access if it was the last expression in the file: it required
changing the `EXPECT` macro to not return if lexing was done.

We also were able to separate the `ParsingContext` into a
`ParsingContext` and `ParsingStack`. This allows, for example,
function calls to not create a new scope/context, but to use
continuation within the parser to avoid recursive calls.

Before, we just created empty scopes for function calls, which is
definitely not the right thing to do, haha.

Add a very small language reference that is very far from even having
begun.

Parse command line options and flags to change output filepath
generated by codegen and the format of the output file.