Upgrade lexer to handle operators made up of multiple delimiters
The lexer is quite simple and general, and has a hard time dealing
with multiple delimiters in a row. This is needed so that we can
parse each ")" or something like that as it's own thing, even when
not separated by whitespace. However, it makes lexing "<<" and ">>"
a challenge, since those are also delimiters.