Codegen: Actually use `ir_..._call()` functions!
Turns out we were creating the call instruction manually, and so
adding the use marking to those functions didn't actually do
anything. Until now, that is, as they are used once more. This also
simplifies things quite a bit in function call codegen, so it seems
like a good change all around.