[Sema] Ensure overloaded functions have matching return type
Also update the test to actually resolve the function `foo`, since
that is when the error is caught. We may want to add another pass
after main sema that goes through all IRFunctions and ensures this,
instead of the way we do it now. This would allow even unused
overloaded functions to be type checked properly.