[Codegen/x86_64] Emit `.byte` directives for static integer init
While this works on my system, we need to eventually make it so that
either the user can give a command line flag to determine endianness,
or it is auto-detected with some preprocessor directive or something.

Code for opposite endianness included, just commented out. Hopefully
if anybody really has problems with it they can find this and see that
they just have to switch which version is used.