[Docs] Update PDFs
It felt kind of weird having my name on the front of everything when
this is more of a community-driven project; so that has been reflected
in the author position of the PDFs, now.

Also make lang name consistent between docs as well as fix some page
breaking issues in `ebook` sizing.