[RA] Coalescence is hard, mmkay?
As evident by `tst/tests/fibonacci.int` returning 1 when it should
return 5 only when coalescing is enabled, the coalescer was
doing *something* wrong. By adding debug logging, I was able to figure
out that copies related to phi's were being removed when they
shouldn't be by the "neither from nor to are precoloured" branch. So
I've disabled that code flow, instead keeping those copies. This has
fixed `fibonacci.int` test failing; it returns `5`, as it should, when
compiled with this patch applied. Logically, it *does* seem like we
should be able to eliminate copies of this nature, so it may be that
we aren't checking enough/the correct conditions.