[x86_64] Some helpers
`regbytes_from_size()` to get the byte size as a number from a
register size.

`femit_reg_to_offset_name()` in order to eventually support stores at
a static address (label) within a struct/array, we need to be able to
add an offset to the name. This allows for that :)