[Codegen] Declarations no longer load their value
I'm not sure declarations should really be doing much of anything,
other than making sure that the thing they are declaring has been
codegenned. Before, we were loading the address, causing, for example,
a declaration of a struct-typed variable to try to load that entire
struct for no reason.

NOTE: This very well may be incorrect, but tests still pass so we'll
see what happens lol.