[x86_64] Mock out lowering of larger-than-register types
I feel like an idiot. For some reason this is so hard for me to think
and reason about. I know the final assembly I want, but how to get
there by having to replace, remove, and move ir instructions around is
not something I am good at, at all.