[Codegen] Variable references now copy the address
This is simply to make the x86_64 backend able to treat all "store"s
of ALLOCA value as storing the value it allocated, not the pointer. In
the future, we should have a better way to deal with this: maybe a
custom backend flags member that x86_64 backend could take advantage
of each time it replaces the value of a "store" with an alloca and
expects it to copy the underlying data, instead of simply relying on
the alloca type. Either way, things work now!