[Tests] CTest compatible test runner
You can now do
```sh
ctest --test-dir bld --output-on-failure
```
to run all of the tests. With the `-j` option, you can even now run
them in parallel. This is a huge speedup! While this test runner is
still very much in it's infancy, I think it's worth it to include.

The algol test runner is still available under the `algol-suite` CMake
target.