¡ `;` as expression delimiter, line comments as `;;` !
Well, this sure was a treat! I owe my lack of carpal to tunnel to
Emacs' kmacro package.

In this commit:

Expected test return status and stdout is now after the new line
comment syntax. Both test runners have been updated (Algol68G & C++),
so all tests that were passing, are passing.

New test for using `;` as an expression delimiter. This is near
necessary when dealing with complex expressions, as the parser gets
confused easily with paren expressions whether or not the thing is a
call with arguments or what. This allows the programmer to separate
expressions easily, and keeps the semi-colon style of comments that I
write anyway.

Allow expression delimiter at EOF; before it would error.

Update the tests README to include details on how to run the C++ test
runner through CTest.

Update the tests README to the new line-comment syntax.