[x86_64] Fix bug in register allocation regarding `idiv` instruction
Division on computers sucks. Division on x86_64 computers sucks even
more. The dividend is stored in the 16-byte rdx:rax register pair,
which means if the divisor is attempted to be stored in either of
those, it will be overwritten either by the dividend itself or by
converting the dividend into an oct-word (16 bytes). This both ensures
programs with registers allocated in this way will not compile
successfully, as well as provides a fix that makes sure any divisor is
never assigned those registers.