[x86_64] Fix type issue regarding loading of statics
IR_STATIC_REF's type is a pointer to it's underlying data; in order to
get the register size needed to store the underlying data, we need to
drop one level of pointer indirection before getting the size. This
only worked before because `integer` is 8 bytes, and so is a pointer.