[Codegen] A (very) basic machine IR
This is definitely "ill-formed" and will require lots of cleanup, but
it is a step in the right direction. "Instruction selection" now
becomes it's own pass which we can easily expand upon and make
better. We are still doing RA on the regular IR, which is a bit
scuffed, so we'll have to change that at some point. Overall, this
helps to share more code between each target that supports the x86_64
architecture.

Again, this is just a very basic, beginner-level MIR that we will
make better over time.