[MIR] Lots of changes
We are slowly working towards having an MIR that fully represents the
program while at the same time being closer in representation to the
underlying ISA that we will be emitting code for. From here, there's
lots of complex stuff to do, like moving register allocation to MIR
instead of IR. And also pattern matching for instruction selection
using a DSL or something would be glorious.