[x86_64/ISel] Use block operands instead of name for conditional jumps
Slowly we are relying less and less on the "origin" IR pointers for
actual functionality. Poggers.

BTW we can always revert this, since the machine instructions don't
care whether it's a block or not and that might needlessly complicate
final code emission. Either way we could now do that by taking
advantage of the actual block operands instead of subverting them.