Minor Bugfixes
RA had an issue with invalid format specifiers since the vreg size
inclusion refactor.

Some fallthrough attributes weren't followed by a semi-colon, and
that's a bad thing ig.

Size of source register is properly set when instruction-selecting a
store instruction (when storing into a static ref). I'm still not sure
how we are going to handle type-based size information in the ISel DSL.