[MIR] PHIs require immediates to be emitted as instructions
We remove them after lowering out PHIs from the MIR, but to do that
lowering we require immediates to have a corresponding machine
instruction. The same issue may arise if an IR_REGISTER is the
argument of a phi node, so we may want to *always* generate at least
one general MIR instruction from each and every IR instruction.