[x86_64] Utilise ISel, begin to cleanup
Get rid of old isel ways: implementing patterns in code
absolutely *sucks*, and I'm glad to never have to do it again.

I still want to remove the `select_instructions` behemoth, but I want
to retain that to build isel patterns off of until the ISel is up to
snuff with what it was before.