[ISel] Even more progress
We've added register-type values to the isel environments, that way
backends can reference registers directly if need be.

We still need some way to actually specificy for each MIROperand
whether or not we should be comparing the value to the pattern's
value.

Fixed broken sorting algorithm (use `j` instead of `i` hehe).