[Codegen] Refactor target specification
I'm still not 100% happy with the system; ideally, a user would be
able to specify all three in one go, like with gcc/llvm target
triples.

NOTE: This very well may not compile; I've changed a lot of things and
may not have been perfect in unweaving the different changes.