[Examples] There are now less examples
1. A lot off the old examples were outdated, and didn't even compile
with the current version of Intercept. It's quite backwards to include
invalid code as an example.

2. By having fewer examples, we can keep them all
maintained/correct/compiling, for the most part. This means people
learning the language won't have to weed through old, outdated
Intercept code just to figure out how it currently works. Ideally, the
examples should always reflect the current state of Intercept, as a
language.

3. I will be working more on the TUTORIAL document, and that will go
hand in hand with figuring out A. some good examples for showcasing
real-world usecases of Intercept and B. other ones that are simple
enough for someone unfamiliar with the language to grasp them easily.