[x86_64/ISel] Handle offsets from locals
Sometimes we want to store the eight bytes of a register into the
first eight bytes of a 32 byte local. Instead of requiring a copy of
the local operand, we can instead use the register size when emitting
and allow usage of the local operand directly in an add instruction.