[Opt] Remove old global variable inlining pass
This pass dates back to the very beginnings of the optimiser back
when functions were still global variables. Originally, it was
intended to allow optimising indirect calls to direct calls by
inlining ’global variables that are only stored to once’—aka
functions—so that we could perform a direct call instead of a—
back then highly janky—indirect call.

Since then, the compiler has come a long way, and not only is this
pass no longer necessary, but it is now causing problems wrt loading
globals that don’t fit in a register.

Optimised performance should not be significantly impaired by the
removal of this pass; irrespective of that, however, removing it is

    1. required, as it completely breaks e.g. passing arrays to
       functions; I am unsure whether this has been broken for a
       while—as the test runner itself has been broken for a while
       wrt compiling with optimisations enabled—or whether this has
       to do with the store forwarding and inlining passes I recently
       added, but either way this pass is much less important than
       either of those;

    1. somewhat expected, seeing as global variables are simply less
       performant than local variables and simply need to be treated
       differently—this mere existence of this pass was jank in and
       of itself in this respect—and

    2. hopefully not going to cause long-lasting performance loss,
       as there are plans to make top-level variables local to the
       top-level function instead of global variables; this would
       do away with most of the performance problems that this pass
       could have purported to solve.

In sum, it really is high time to sunset this relic just about now.