[IR] Add IR_POISON instruction
I was going to use this for something, but I ended up not needing
it, but seeing as I’ve already added we might as well keep it around
for later.

A poison value is an invalid value that can be used to replace
values that are to be removed or generally invalid computations.