[Opt] Refactor dominator tree construction.
The way in which we were previously constructing dominator trees
was not only janky, it also produced nonsensical (or at least
categorically unreadable) results.

This commit is a complete rewrite of that using the Lengauer-Tarjan
dominator tree construction algorithm. We are currently not using
them anywhere, but since we can also visualise CFGs and DJ-graphs
(which currently are just dominator trees) now, I’ve taken a good
look at those, and everything seems to be correct now.