[Refactor] Make IR opaque.
This makes all IR structs (chiefly: IRInstruction, IRBlock,
and IRFunction) opaque. This means that their members cannot
be access outside of IR implementation files. This also means
that use marking is now completely automatic.

IR implementation files are now only `ir/ir.c` and `ir/inline.c`.
The inlining code has to have direct access to the structures
because of the weird way in which it creates/copies them. The
implementation of IR structs is now in `ir/ir-impl.h`. Any
parts of the codebase other than the aforementioned `.c` files
should only ever include `ir/ir.h`.

For nearly all instruction types, there are now 2 functions to
create them: the functions starting with `ir_create_` create,
but do not insert, the instruction. The functions starting with
`ir_insert_` create and insert the instruction in the current
insert block of the context. The latter are mainly used during
initial IR generation.