[Intercept/Parser] Make a missing semi-colon a warning
If I could, I would have this still be an error, just not a fatal
one. While if you leave every semi-colon out of a program,
things *will* break, every once and a while one gets left out where
one isn't needed and having that break the whole compilation is really
annoying.

In the future, I would like to implement something more akin to GLR
parsing, where there is an ambiguous AST node for "expression list"
and "function call", and only once the type of the first expression in
that list gets resolved may we resolve the node into a function call
or not.