[Intercept/Parser] Use the right scope for declarations
To allow for top-level local variables and so as to not make every
function a local function, the Intercept parser distinguishes between
a ‘Global’ scope and a ‘TopLevel’ scope, the latter being the scope
of the body of the top-level function.

This commit makes sure that only local top-level variable decls
are placed in the top-level scope, and that every other top-level
decl (variables, functions, aliases, and types) is instead placed
in the global scope where it belongs.

This fixes #87