[LCC/Codegen/ISel] Ætérnal C++ magic
Basically, we will need to like do the whole isel thing where we build
a window of instructions to apply patterns to that is the size of the
longest instruction. To do this, we need to iterate patterns, and,
since those are types, I had to get help to figure out how to do that.

With this (or something like this) technique, we should be able to
iterate the PatternList parameter pack and go until we have matched a
pattern. If we don't match a pattern, we pop off the front, keep
going, the whole nine yards.