[x86_64] Utilise clobbers in ISel, among other things
Take advantage of the new `NewVirtual` operand kind to implement the
storing of an immediate into memory (which is not directly possible in
  one opcode on x86_64).

Add dereferencing moves into the x86_64 machine instructions. Still
not sure how I feel about going about it this way vs operand kind
patterns of a single instruction denoting this sort of thing, but we
are rolling with it for right now.

If there is an attempt to get a string representation of a zero-sized
register, assume it's 64 bits.