[LCC/Codegen/ISel] New operand types
Also, this is a good point to mention I've realised I did this
incredibly stupidly which results in overly verbose pattern
specs. Basically, the "value" struct of `o<N>` and `i<N>` should *be
the operand itself*, hard-coding it's kind. That way, instead of
having to do `Operand<OK::Function, Function<>>` to get a function
operand, `Function<>` is simply a function operand.