[Intercept/IRGen] Don't force pointer decay for alloca arguments
Or something to that effect. If the argument is referenced in the IR
by an alloca, and we expect the type of the thing the alloca
allocated, we can almost assuredly assume what we want is to
dereference that alloca and load from it. This may be an issue in sema
or something missing a cast or lvalue to rvalue conversion, but I'm
/pretty/ sure this is caused simply by how the IR is being generated
and the fact that, in the IR, lvalues are pointers (afaik).