[Opt] Fix bug in InstCombine
When handling `sub`, were were `cast<>`’ing the LHS and RHS to
IntegerConstant, comparing `lhs == rhs` and replacing the sub
with `0` if true.

However, `lhs` and `rhs` will both be `nullptr` if neither is
and IntegerConstant, thus making this check succeed even though
the operands may not be equal.

What was intended here was actually `sub->lhs() == sub->rhs()`.