[LCC/Codegen/MIR] Make utils `inline`
I /would/ move the definitions to an implementation file and just
forward declare them in the header here, HOWEVER, auto return type
deduction doesn't work like that. So, I'd have to manually write out
each instantiation of each template, rather than just being able to
deduce it from the returned expression. That would bloat all of this
by 100s and 100s of lines. To keep it simpler, at the expense of code
size (not something we are even remotely worried about right now), we
choose to just allow multiple definitions of these functions.