[x86_64/ISel] Return global
This is only done in code because I don't want to write an entire
thing into the ISel TMP right now just to access the size of a global.

I should eventually add a `SizeofOperand<OpIndex>` type that tells the
pattern replacer to get the size of the input operand at the given
operand index and write an MOperandImmediate with that value.

I should also eventually add a `When<>` type to the PatternList that
specifies conditions for the pattern to apply, like specific values of
input operands, and more.