[CMake]
No precompiled headers because it keeps filling up my `/tmp`
filesystem completely with `.pch` files that are over 20 megabytes
each.

`PROJECT_SOURCE_DIR` is just stupid to use, it basically just asserts
that this file can't really be used in any other CMake files, and I
don't like that. If we wanted to be that specific about paths, we
would want to specify `CMAKE_CURRENT_LIST_DIR`.