[Glint] Error on Unhygienic Expansion of a Macro
Also fix a very complicated location bug in the lexer regarding a
macro definition immediately followed by an invocation of a macro; in
fact, locations of expanded macro tokens seem to have been completely
wrong before this.

Lots of comment and error message improvements, as well.

Oh yeah, and we weren't handling errors properly at the top level of
parsing; we never set the error flag in the context.