[x86_64/GNU ASM] Use immediate size when storing immediate into local
This is important for cases when the local is an array or anything
like that and we are just trying to store into the memory at that
location. NOTE: This may break certain cases of assigning immediates
to locals IFF both sema and IRGen are completely not doing their job.