[Glint/Sema] If any expr of a block fails, the block errors and we exit
When a preceding expression within a block is known to be invalid, it
stands to reason that the following expressions are not going to be
valid (code within control flow is likely to use and rely upon the
result and/or side effects of this known-bad expression).

Reduces useless, redundant errors in compiler output.