[Glint/LangTest] IR tests
Basically, if you don't access a parameter at the top of a function,
right away, it may be overwritten (due to it possibly being in a
register and register allocation having no real clue about
that). While this is sort of a limitation of the LCC Register
Allocator (at least currently), we force Parameter instructions toward
the top by copying all parameters into a local.

Ideally, we could reference a parameter instruction anywhere in the IR
and it would "automagically just work", but, we live in the real
world, and I'm the only one who wrote the register allocator, so
things definitely aren't perfect.

Hopefully, one day, the IR, MIR, and RA are smart enough to not make
that blunder of overwriting parameters that have not yet been
accessed, or something (or at least having a check for it that errors
out so broken code isn't emitted with no warning).