[LCC] `Location` references file not-yet created
Now that we have a language that creates files /after/ creating
locations from parsing the original, single source file, it was
possible for a location to reference file_id 0, which didn't exist at
the time, and then later that file got referenced, before an error was
emitted. This caused the location to be applied to the wrong file, and
the user would get very funky error messages.