[LCC] "Fix" Defining Uses
I am not entirely sure if this is correct or not, but, without this,
we spend about 8000x the amount of time we should calculating defining
uses. The idea is, whether or not we've visited a block shouldn't
change depending on which control flow we follow, BUT the registers
seen (live values, basically) SHOULD change depending on which control
flow we follow.