[RunTest]
Multiple sources handled... Yay! It's kind of awful, but, it works.

I didn't want to implement an entire dependency resolver or whatever,
so it just goes in opposite order as they were defined in the file,
and treats the "last" (first in file) source as the executable source,
and all the others as dependencies of that last source. That means you
can't really do A imports B imports C. That is, you are reduced to one
"layer" of imports, due to the "simplicity" of the build system.