[LCC/IR] Struct types now respect alignment
It also means they require an alignment upon creation. We could have
made it so that there is an `aligned_struct` sort of type that would
have it set specifically, but most languages would just end up using
that one anyway.

TODO: Still need to update IR parser with syntax to set alignment
explicitly, and also emit that when we print the IR, if need be.