[Glint] Syntax Change
To reduce parentheses in common cases (a direct design goal of Glint),
I have decided to remove `&` as a binary operator. For consistency, I
altered all of the bitwise binary operators to `bitand`, `bitor`,
`bitxor`, and `bitnot`. I'm not super happy about how it looks, but,
how often are you really doing bitwise operations /in userspace/ these
days? Maybe it doesn't need the looks, and the reduced parentheses in
every addressof call argument may just make it completely worth it.