[Glint] AST Serialisation
Basically, Glint's BMI now contains an expression table.

In depth: Glint has unique module semantics that aren't really
explored by current popular language implementations like GCC's or
LLVM's C++. For example, you may import a module that exports
templates. Now, this may be odd to you, since importing something
usually just tells the language to treat everything that module
defines as if it will be linked with and it all already exists. A
template, however, is already removed at the time of linking. So, the
old C style approach of "anything imported is something that may be
later found in the executable, resolved by the linker" is wrong. The
Glint approach is "anything imported is defined elsewhere". The Glint
module system actually has no "knowledge" of the linker or that entire
system of building; modules merely guarantee that exported objects are
accessible upon import, and need not be (re)defined. To make this
work, template definitions (and templated function bodies and
templated struct bodies) are all actually serialised (yes, the AST
itself), and later re-built upon deserialisation.