Remove Examples
Because they aren't a part of any test framework or anything like
that, and the language is quickly growing and changing at the moment,
these are mostly just examples of an out-of-date language that doesn't
even exist anymore. If you want to see Glint code, read the book, read
real programs (like GlintSDL3 example, or the standard library), or
read tests (i.e. `lib/glint/tests/corpus/`, `runtest/corpus/glint/`).