[LCC/Diags] Introduce `Fixes` Feature
Now, you are able to let the user know that there is a set of
operations they may apply to make this diagnostic no longer be
emitted.

It'd be cool if we could also emit `diff` files or `ed` scripts or
`sed`/`awk`/`perl` commands to actually apply the edits that the fixes
describe.

Eventually, it'd be cool if the driver could ask you what you actually
meant, given some options (i.e. you typed 'foor', did you mean 'foo'
or 'four'?).